requirements.txt
keyboard
PIL
pytesseract
pythoncom
pyHook
requests
baidu-aip
pymouse